gnrc_netapi: make doc a little more consistent
Throughout the documentation `gnrc_nettype_t` is referred to as a
"protocol type". The documentation for the functions
`gnrc_netapi_dispatch_*()` that were added later only refer to it as
"type".

This fix changes those to "protocol type" to make this clearer in the
context of these functions and unify it with the rest of the
documentation.